Move Uniconfig API and types to dashboard
