+++ packages/frinx-dashboard/src/helpers/types/uniconfig-types.ts
